fix: decimal lifetimeValue

+++ Core/Projections/CustomerDash.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string CustomerName { get; set; }
         public int TotalOrders { get; set; }
-        public int LifetimeValue { get; set; }
+        public decimal LifetimeValue { get; set; }
         public IEnumerable<RecentOrder> RecentOrders { get; set; }
     }
 }
